requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.9.0
openai>=1.30.0
python-dotenv>=1.0.0
websockets>=13.0.0
httpx>=0.27.0
python-multipart>=0.0.6
aiofiles>=23.2.1
pydub>=0.25.1
chromadb>=0.4.0
PyPDF2>=3.0.0
python-docx>=1.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
